Tweaked ReplayGain support to handle mp3gain properly. Since mp3gain modifies the MP3 files themselves rather than simply adding a tag, audio encoders shouldn't apply it by default. So, I've added a lossless_replay_gain() method to AudioFile which indicates whether the default for adding ReplayGain should be on or off - assuming ReplayGain is available at all.
Also, with this lossless/lossy ReplayGain intrastructure in place,
it should be possible to apply ReplayGain to *any* sort of audio file
by piping the raw PCM input data through a ReplayGain calculator and then
rebuilding the files via from_pcm() with modified values.
However, this calculation is too slow for pure Python to handle
so an external tool (wavegain?) or yet another Python C module will
be necessary.