Converted tracksplit to have unified TOC/CUE support. Still need to update its documention, however. I've not yet decided what to name these things in the docs and command-line.
Having both --toc and --cue options is pointless duplication
(the program can figure out which is which, and no one's going to use both).

Inventing a --sheet option is a bit more accurate, but it's inventing
terminology that one won't recognize at first glance.

So, having the option as --cue but allowing both CUEs and TOCs is probably
the most ergonomic way of handling them since most people will have
.cue files.  But, that feature will need to be documented.