Preliminary conversion of M4A to optionally use Nero's Linux codec. This required fixing a critical error in which rewriting 'meta' atoms messed up the whole file in instances where 'mdat' is the final atom.
'meta' is the file's metadata and 'mdat' is the file's actual data.
faac placed 'mdat' in the middle of the file and 'meta' at the end of the file.
So in that case, rewriting 'meta' had no effect on 'mdat's position.
neroAacEnc switches these two around.  Unfortunately, the 'stco' atom
contains file offset values which are broken when 'mdat' moves.
'stco' is now rewritten with proper offsets when 'meta' is adjusted,
fixing this error.