Very preliminary support for audio format decoders implemented in C.
The purpose if this is twofold.

Primarily, this reimplementation is an instructional exercise meant to
demonstrate format decoding in a simple, easy-to-understand fashion.
Secondarily, these built-in decoders can cut down on the use of external
binaries.

Although implementing audio decoders in Python works well for the former,
it's far too slow for the latter - in spite of any attempts at optimiziation.
I've tried to use other high-level languages for this task
as a sort of sister project, but none of them are fast enough.

If successful, I'll move on to format encoders later on.