Add coffset documentation to Shorten chapter.
Though a bit of a headache in source form, it's not too awful
once turned into proper math formulas.
I've omitted details about bit shifts, but that may be a case
of using a little inaccuracy to save a lot of explanation -
especially since the reference encoder doesn't like to generate them.