Add SHNDecoder sample_rate and channel_mask attribute setters.
The idea is that when someone calls .to_pcm() from Python,
we'll build a SHNDecoder, parse the file's built-in headers
and use that data to fill-in .sample_rate and .channel_mask
before sending it out.

Because it's such a big, dirty job, I'd rather do that work from Python
(which has a bunch of pyconstruct parsers I can use) rather than
from the low-level C stuff which is meant for speed.