Add wasted-bits support to ALAC decoder, necessary for 24bps decoding.
Curiously, this is one place ffmpeg gets it wrong
by bit shifting the result up regardless of whether wasted bits were added.
Not that I'd hold it against them since 24bps audio is so rare in practice,
but I was surprised to find my own intuition on the subject to be more
accurate than their code.

All that remains for the decoder is additional testing
and another round of format documentation.