Update ALAC encoder to handle random noise.
Unlike FLAC or Shorten, which can build up to 32-bits of residual value
(very inefficiently) in extreme cases, ALAC residuals have a limit
of the input's bits-per-sample, roughly.
This means that an attempted residual of random noise will overflow
its escape code, so I've added a "RESIDUAL_OVERFLOW" exception
which instructs the encoder to abandon compression altogether for that frame.