Fix a regression in ALAC stco atom generation.
Bouncing mdat creation to a temporary file causes all the offsets
to start near 0, which won't be right come full file creation time.
Now we count the sizes of the pre-mdat atoms, adjust the stco offsets,
then build a real set for the actual file.

FLAC got this right by basing the seektable from the start of audio data.
Making it an absolute offset is nothing but headache.