Add a set of TODO items for the long-overdue broken file handling.
It's been low on the priority list because many formats were punted
to helper programs (which have to catch broken files on their own)
and really broken files simply aren't found in the wild that often.
But now that I'm handling more and more formats internally,
it's best to be more picky about problems.