Add some preliminary infrastructure for bitstream error handling.
We now read and propogate ints instead of unsigned ints,
which is necessary for catching EOF.
Now, if EOF is encountered, we simply dump an error and exit.
This will be fleshed-out with a primitive exception handling stack
so that fatal read errors are pushed up to Python so it can build
friendly error messages.