Preliminary MP3/MP2 verify() method.
This only checks that one can walk through the file
without errors and that the headers are consistent throughout.
I should probably check an MP3's CRC-16, if present,
but it's so rare and covers so little of the file (perhaps 10% of each frame)
that there might not be much point to it.