Convert C Bitstream struct to have embedded read methods.
Analagous to the bitstream writer's built-in writing methods.
In this case, we'll specify big-endian/little-endian at bs_open()-time
and it'll set the appropriate read functions without us having to
remember to call read_bits_le(), read_bits_be() and so on.

The bitstream writer will require something similar
for its write_*_actual functions, but all that infrastructure
is already in place which will limit conversion to updating
its bs_open() call.