Add some preliminary WavPack bitstream decoding documentation.
It raises the bar for documentation inelegance.
I've tried simplifying and "de-optimizing" the process without success.
"holding_one" and "holding_zero" act much like bits pushed back onto the
stream, but that's not quite right (determining which bit to un-read
depends on the unary value *above* the previously un-read bit).

It's frustrating because while the purpose of steps 2 and 3 of WavPack's
residual decoding are relatively clear and consistent with its
theory documentation, the purpose of all the bit twiddling in its
unary-reading stage is not obvious and not explained anywhere I can find.

I'll redo the documentation as needed should I sort it out better.