Add preliminary false stereo support to WavPack decoder.
As usual, the false_stereo.wv test case doesn't yet work
because the decoding process flies off the rails around block 6
for the duration of that block.  It's probably not a problem with
the actual false stereo handling (since there's not much that
can go wrong with it) and more likely some decorrelation pass gone
awry - again - since the actual decoded bitstream block seems okay.

Odds are it's just another stupid bug.  But the sheer number of
stupid bugs illustrates how much more complex WavPack is
compared to FLAC or ALAC.