Add some error checking against hybrid WavPack files.
I'm punting on WavPack's hybrid mode for now.
It's an odd fit for Python Audio Tools' structure
since a WavPack's losslessness would depend on whether
there's a correction file somewhere in the vicinity of the
main file, so WavPackAudio's __init__ method would have to do some hunting.

It also adds a significant amount of complexity for a feature
that I've never seen used in the wild - despite its tagline.

So, hybrid mode will wait until version 2.17 while I build
the encoder and move on to more user-centric things for awhile.