Finally fix the WavPack residual state generation mechanism.
It's a bit bigger than the reference encoder's implementation,
but should be a lot easier to understand.  The reference WavPack
encoder mixes the bitstream handling routines, unary value
generation and holding_one/holding_zero state into a couple of
big routines.  I've chosen to split up that functionality
into more functions with a lot of comments in the hope that
it'll be easier to understand and likely to be correct.

Though escaped unary is still pending, things are looking up.