Add more assertions and bugfixes to WavPack encoder.
However, it looks like the value wrapping will need an overhaul.
Blocks, even across the same set of output channels, may not have
the same number of channels when false_stereo is taken into account.
Not to mention that decorrelation passes with negative terms
(such as produced by 10 and 18) can't be applied to 1 channel blocks either.

The simplest solution will be to dump wrapped values entirely
and start with a new set should the channel counts differ.