Critical fix for WavPack encoder's internal RIFF header generator.
Adjust some of the channel mask handling tests for WavPack.
Given an undefined channel mask, the previous behavior was to assign
them in .wav order.  The new behavior leaves the mask blank,
which may make some wav decoders unhappy.
I'll leave it as-is for now until I've done more testing.