Update editxmcd to Urwid-based text interface.
Since most of the tools are already text-based,
it makes sense to minimize the amount of graphical tools whenever
possible.  For example, a text-based editxmcd runs over ssh and so forth.
And, as long as I've got the requirement in place, I may build
other interactive Urwid-based tools as appropriate.

This also neatly removes the GTK requirement, except for coverview.