Add foreign AIFF chunk support to FLAC.
Still a bit of a mess, though.
My ultimate plan is to pull the wave-specific methods out of AudioFile
and have the convert() method use them as appropriate from subclasses
depending on what's being converted from and to.
This avoids building a ton of special cases into the base class.