Finish up convert() method conversion by adding documentation.
This essentially means that an audio file potentially has 3 levels of data:

1) the raw PCM data handled by to_pcm()/from_pcm()
2) side data like non-audio RIFF chunks handled by convert()
3) human-centric data handled by get_metadata()/set_metadata()

It's tempting to combine levels 2 and 3 into one.
For instance, WaveAudio could return a RIFFChunks MetaData object
which would place non-audio RIFF chunks with a set_metadata() call.

The reason that approach is a Bad Idea is because WavPack and Shorten
need to have those chunks - which we won't be editing anyway - at encode-time.
Otherwise, we'll need to *rewrite the whole file* in order to add them later.
So, short of combining the encoding/metadata setting into a single step
(which would require a method like convert() anyway), this is probably
the best way of handling that data for now.