Add preliminary dvda2track and dvda2xmcd utilities.
Still need to nail down the exact interface, however.
It'd be nice to use something like the -c/--cdrom option,
only with a mount point, to have as a default for finding the AUDIO_TS
directory.  Unfortunately, this breaks down on places like Mac OS X,
where the DVD's name becomes part of its path.

So, I'll need to think of a short "source-dir" style argument,
to differentiate it from "-d" which is for output directories only.