Several adjustments to handle a different style of DVD-A extraction.
Although we know a track's start and end sectors,
there's a good possibility the track's data doesn't begin
at *exactly* the start sector given; a given sector
may contain the end of one track and beginning of another - it seems.

A more reliable approach may to treat an entire DVD-A title
as a PCM/MLP stream and pull it apart with the pcm_split() function
as needed.