Add preliminary Controller/Player objects for interactive trackplay/cdplay
The idea is that the player runs in its own subprocess
and communicates with the controller by passing pickles back-and-forth
across a pair of pipes.

Player can be subclassed into OSSPlayer/PulsePlayer,
or any other sort of player so long as the interface works.
Controller can be subclassed into non-interactive/Urwid/etc.