Re-support cdrom_read_offset via CDDA class.
The hideous hack of reading all the CD data at init-time is avoided
by combining a conceptually simple PCMReaderWindow
(which truncates or extends a PCMReader as needed)
with an adjustment of the CDTrackReader's start and end sectors as needed.

This ensures that CDs with data at the very beginning and end of tracks
will not lose anything when used in conjunction with the read offset -
except data at the very beginning or end of the disc, naturally.