Add a more generalized progress managing queue.
This queue takes functions and arguments, executes them in parallel
and keeps track of pickled results.   The difference is that
the queued functions must have an additional callable "progress" argument
which the queue's subprocess will wrap accordingly so that
calls from child jobs propogate to the parent's ProgressDisplay.

Exception handling is still a bit preliminary, but this queue should
be usable for track2track, tracksplit, trackcmp and trackvery
once it's fully functional and tested.