Add progress display to tracksplit.
I've removed its multiprocess operation in the process.
Though slightly faster with multiple cores, tracksplit's multi-CPU
operation is both more complicated and nonintuitive to watch.
The period in which it dumps the input file into a bunch of temporary
files (so that the temp files can be encoded in parallel)
makes one wonder what it's doing while it uses a ton of CPU
without putting anything in the output directory.

The simpler, slightly slower approach better follows the principle
of least surprise.