Add preliminary man page generator.
The idea is that the pages themselves are stored as abstract data files
and the troff data is generated as necessary at install-time.
Though the data files are yet more XML (so Python can handle them easier),
this should ensure that the man page output is more consistent
than it is now.
Additionally, I'll be able to generate man pages in other formats,
like reStructuredText for use on the website.  Although troff has html
targets now, they're not especially pretty or flexible.
This approach does a better job of separating content from presentation.