Add dynamic Huffman jump table generator.
Though still in need of cleanup, its function is to take a list of
frequencies and create a table suitable for use by bs->read_huffman_code
on-the-fly rather than as something prebuilt in a .h file.

Although prebuilding the tables is most convenient
(since there's no need to anchor them in memory anywhere, or deallocate them)
formats like Vorbis don't define them ahead of time, which makes that
impossible.  Thus, the necessity of this runtime approach.