Add preliminary Ogg Vorbis decoder.
The point of this is to implement a lossy decoder that's already
well documented (though I'll try to translate the official spec
into something that's more my style) and then adapt its principles
into other lossy codecs like Musepack, MP2, MP3, etc.

The long term goal is to cut down the number of external binaries
to the bare minimum.