Add another big pile of Vorbis decoding documention.
The official spec leaves the filling of the Huffman table to a single diagram
which makes it harder than necessary to figure out.
Diagramming out the codebook parsing is also a little easier to follow,
though it could do with some cleaning up
(and perhaps I can integrate the ordered entries if I give it some thought).