Add placeholder support to BitstreamWriter.
This is to handle fixed-size values which must be written out-of-order.
For example, the total samples field in FLAC's STREAMINFO block,
Ogg page checksums, total size field in WavPack block headers and so forth.
The idea is that we write dummy values to the stream, store their positions
in a FIFO, and then fill them in on a second pass, such as with:

stream = bs_open(fp, BS_BIG_ENDIAN);
stream->write(stream, 8, 1);             /*normal first byte*/
stream->write_placeholder(stream, 8, 0); /*second byte is 0 for now*/
stream->write(stream, 8, 3);             /*normal third byte*/
stream->fill_placeholder(stream, 8, 2);  /*second byte is rewritten as 2*/