Adjust the BitstreamWriter placeholder routines for out-of-order calls.
Instead of requiring FIFO-style fills, they can now be done in any order.
This could be useful for formats like ALAC in which we'll need to
calculate and generate atom sizes in a more stack-like fashion.

Need to build write_placeholders()/fill_placeholders() routines
which handle creating several placeholders at once via a format string.
For example, FLAC's STREAMINFO has several placeholder fields clustered
together.  Not only do we not want to bounce back and forth with fsetpos
for each one, but the fill_placeholder routines won't function
as expected on file output that's not byte-aligned (the preserved
output stream state in a placeholder can't be updated at fill-time).

Supporting a cluster of output solves both of these problems.