Convert ALAC/M4A to use the new M4ATaggedAudio routines for metadata.
Still a bit rough, but it does pass all the unit tests.

The hdlr atom is particularly interesting.
Some encoders stick padding on the end, while others do not.
In the future, I'll need to store that extra cruft and
recreate it at build time in order to fully round-trip properly.