Lots of updates to ID3v2.3 documentation.
Though I'd initially intended to document the whole thing,
a lot of tags are sparsely documented, of limited use or
aren't used in the wild anymore - which makes them hard to test.
Rather than spend lots of time on dark corners of the spec,
I'll concentrate on the frames people actually use.