Preliminary rebuild of ALAC documentation.
When it's finished, one should be able to build a proper decoder
just by walking through the documentation and coding its steps.
I believe FLAC is at that point already, but ALAC has a lot more
"gotchas" which will take more work to turn into something that's
easier to follow.