Secondary rebuild of ALAC residual decoding documentation.
Still frustratingly rough.  The trouble is that numbered steps
are good for illustrating formulas, but lousy at flow control
whereas flowcharts are good for if/else/goto flow control,
but ill-suited for complex math.

I'm going to need a better way to illustrate how these functions operate
in order to document them in a reproduceable way.