Preliminary rebuild of FLAC LP coefficient calculation documentation.
Even cut down into pseusocode, this needs lots of additional commentary
to explain exactly what it's doing because it performs calculations
on lots of list items in non-obvious ways.
It's actually a little easier to follow from example because one can
see which values are being combined with which.

This process is actually exposing bugs in my FLAC implementation,
which makes it even more useful than I'd first hoped.