Finish up FLAC encoding documentation.
The nature of being an asymmetric codec is that the decoding documentation
is relatively short and easy to follow, while the encoding documentation
(especially the LPC subframe algorithms) stretches on for pages and pages.
However, in theory, one should be able to implement a whole encoder
from the documentation alone.