Overhaul FLAC's FIXED subframe order documentation along with implementation in Python encoder in order to be made much simpler.
Originally, I'd implemented the FIXED subframe generator in two steps.
The first step took a bunch of subframe samples and some arbitrary order
before building a FIXED subframe in that order.
The second step took a bunch of subframe samples
and tried to figure out the best FIXED subframe order to use.

However, figuring out which order to use actually involves
calculating the residuals for each order and using the order
with the smallest absolute value sum.

So by combining two steps into one,
the algorithm becomes more elegant and easier to follow.