Add rebuild of the low-level array handling routines.
The idea is to replace the messy, ad-hoc collection of ia_*/fa_*/iaa_*
functions with ones with a consistent interface across
the various styles of arrays.

These functions tend toward array-at-a-time operations.
For item-at-a-time calculations, it'll be best to resize the internal
C array, pull it out and assign a bunch of items as needed.  Like:

function(array_i *array, ...) {
  int *data;
  int i;
  array->resize(array, sample_count);
  data = array->data;
  for (i = 0; i < sample_count; i++)
     data[i] = some calculation here
  array->size = sample_count;
}

This style removes most of the overhead for high speed loops,
while making arrays easy to deal with the rest of the time.