Bump to version 2.18alpha4
WavPack is Done Enough for now.
I'm glad to have it documented more thoroughly,
but overhauling it took far longer than I anticipated.

I can now get back to all the Python-level TODO items I'd been neglecting.

From there, the roadmap is to overhaul Shorten (which won't take long)
and DVD-A (which only has a decode portion) before moving on to a 2.18beta.