Convert function-based Bitstreams to use existing buffer objects.
This makes them more conductive to building substreams with.
For example, br_open_external's read() function could be
implemented with a BitstreamReader of its own that dumps data
to a BitstreamRecorder, whose internal buffer is then copied as output.

The whole process is much like a BitstreamReader substream done in reverse
which may be more appropriate in some codecs where inner data may
span a container's blocks in more unusual ways than Ogg FLAC or WavPack.