Add basic DVD-A structure parsing as a preliminary C module.
Once I build this as something that works standalone,
I'll hook it into Python as a set of objects
that can generate working PCMReaders.

This is where object arrays come into play
since if there's a parsing error, I need to be able to deallocate
only the sections of the overall structure that have been
successfully built before eventually erroring out.