Rebuild AOB stream decoder with different design and working PCM decoding.
The previous version assumed the whole title was identically formatted
across tracks and had AOBPCMDecoder or MLPDecoder pull packet payloads
from a Python reader and then split the tracks - essentially a "pull" model.

The new version appends packet payloads to a FIFO and decodes as
many as possible with PCM or MLP, depending on which is detected
at track init-time - more of a "push" model.

The latter should better cope if tracks change format midway through a title.