Fix a WavPack wave footer encoding bug.
Ensure WavPack encoder doesn't try to build an oversized wave header.
That is, if the size of a hypothetical "data" chunk size field
is too big, forego building the header and keep it as a DUMMY sub-block.

In practice, only old WavPack decoders demand the wave header sub block
so this shouldn't present much of a problem.