Convert .wav, .aiff and .flac to use the new chunk interface.
The idea is that chunks implement a handful of methods
so that a .wav or .aiff file can be built from them.
This allows very large chunks (like wav "data" chunks)
to be kept on disk while very small ones are stored in memory
without file generators having to know the difference.

As a side benefit, FLAC can use the same interface
to avoid storing all its PCM data in memory.