Convert progress update to polling-based system.
The upsides to this interface are that the progress display
polls its children at regular intervals rather than
sorting through lots and lots of child output
(most of which gets thrown away),
and that it's a lot simpler to follow.

In practice, it flickers like crazy because the child processes
don't respond to the polling fast enough so it blocks
with an un-updated screen until everyone catches up
and that sucks.