Optimize substream_append calls when using function-based BitstreamReader.
Instead of making an ext_getc call for each byte
(some of which will call the external function)
this fills up the buffer as needed with multiple external function calls
and uses memcpy to populate the substream.

This should be more efficient, especially for the common case
where the buffer already has enough data for a substream.