Convert PCMReader's read() method to use pcm_frames instead of bytes.
The bytes argument is a relic from when codecs were implemented
by piping data from one external program to another external program.

Now that most codecs are implemented internally
(or going in that direction), it makes more sense to have the argument
a whole number of PCM frames instead of converting a byte count
back and forth.