Convert utilities and their unit tests to use Filename object.
A Filename is an immutable object representing a file
which may or may not be on disk.

It can be converted to strings or unicode objects
(removing the need for the old Messenger.filename classmethod)

But most importantly, it contains stat(2)-derived dev and inode integers
so that two Filenames with different paths but pointing to the
same actual file will always hash and compare equally.

This is crucial for ensuring that utilities don't accidentally
overwrite input files with output files during operation.