Update ID3v2 and M4A to treat track number values of 0 as None
For example, ID3v3 stores track number and track total in the same field.
If track total is present but track number has a placeholder value of 0,
treat the 0 as None.

M4A works similarly, but the track total field must always be present.
So track total is always treated as None instead of 0.