Add preliminary True Audio documentation.
I may add .tta support and bump Monkey's Audio to some future version.
Neither is very common in the wild but True Audio seems better assembled
and can use a lot of existing infrastructure.

Like Shorten, it may be a case of adding it because it's very easy
rather than due to any demand.