Fix and document a flaw in True Audio's hybrid filter.
Having to wrap the hybrid filter's sum into a 32-bit signed integer
looks ridiculous in Python and in the LaTeX format documentation,
but that's what's required in order for the format to work "correctly"
(but at the cost of larger files in those instances).

Shorten has a similar deficiency in its LPC commands,
but at least there I can decide not to use them.
True Audio's filter isn't optional.

Also fix a bug in True Audio block size calculation.