Change names of array structs to something easier.
This involves changing an array of ints from "array_i" to "a_int"
which puts the contained type in the name and the level
of arrays in the prefix.
So an array of arrays of ints [[1,2,3],...] becomes "aa_int", etc.

It should make the code a bit easier to follow, I hope.