Add multiplier capability to bitstream's parse/build routines.
That is, if one wants 100 signed bytes, you can specify "100*8s".
This should cut down on cases of having to do .parse("8s" * 100)
at the Python level in order to fetch several values at once.

The .build() routines also take an iterator of values to be
output rather than something that needs an index.