Minor low-level buffer type handling cleanups.
In theory, this makes it very easy to convert window_start/_end variables
to pointers which is a closer match to what the buffer window is
and would limit maximum buffer size to whatever a system's RAM could hold.

In practice, switching offsets from unsigned integers to pointers
incurs a significant speed penalty, likely due to all that 64-bit math.