Add volume adjusting widget to trackplay's GUI.
This widget can be extended later to allow switching audio output.

The downside is that volume adjustment during playback is sluggish
since the player's command queue is lower priority than outputting audio
and both are processed in the same thread in the player subprocess.
Still pondering the best way to handle this.