Add pseudocode formatting language to reference documentation.
Convert reference documentation Makefile to gmake
which sacrifices the portability of simple suffix rules
for more powerful ones which can compile the pseudocode compiler.

The pseudocode compiler takes a file and compiles it to LaTeX blocks
for use with the algorithm2e LaTeX package.  The language is simpler
than writing LaTeX directly and smart enough to format common
structures correctly (like multiple variable assignment).

This may seem like a lot of work to cut down on a little LaTeX.
But by adding another pseudocode output format down the line,
along with having the bitdiagrab/bitparse/byteparse utilities output SVG,
the whole mess of reference documentation can be ported to HTML.