Update bitstream module's internal buffer sizes.
This is useful to minimize the number of calls to the C bitstream
functions by trying to cover as much of an unsigned value as possible.

Although it's straightforward on the BitstreamReader side
by using Py_BuildValue to parse an unsigned int,
it's trickier on the BitstreamWriter side since bits need to fit
into an unsigned int *and* long value (which may be the same size)
without overflow problems.