Dynamically choose audiotools.cfg location on install
It is unsafe to assume that /etc will be writable by the installing
user. For users that are deploying with virtualenv this poses a
problem as they may not be running as root. This change check to see if
the installing user has access to the configuration directories and
selects an appropriate directory if it exists.