Update mp3lame library probing heuristic.
Since's the mp3lame doesn't seem to advertise itself in pkg-config,
there's few easy ways to detect it.  So instead of forcing users
to specify it in the config file, we'll just assume that the presence
of a working lame binary implies its library is present also
(which works if the user installed from source).