Wrap etry and pop_callback functions in macros.
This allows them to display the filename and line number in the warning
when popping from an empty stack.

Although it would be nice to have the bitstream methods display that
also (e.g. unmark(), rewind(), free()), that's a bit too much hacking
for something that should never be seen in normal operation.