Add AccurateRip V2 support.
Unlike V1, this does *not* work on over a range of samples with offset detection;
it only verifies the stream at offset 0 and if that doesn't match
utilities fall back to checking version 1 and offsets.

Though the math seems simple enough, I've yet to deduce a way to
calculate checksum deltas needed for sliding the "window" in constant time,
so this will have to do.

Simplify a lot of the AccurateRip routines in the process to
leave fewer places for bugs to hide.