Add seeking to low-level BitstreamReader module
Analagous to fseek, this allows seeking to some arbitrary byte
in the stream relative to the start, current position, or end of the stream
by calling fseek or a seek() function as needed.

This is *not* a substitute for the mark stack since the reader's
state and buffer contents are not preserved after a seek.