Add context manager support to BitstreamReader/BitstreamWriter
That is, performing:

>>> with BitstreamReader(open(file, "rb"), False) as reader:
>>>     <do stuff>

will call reader.close() on exit, which closes the opened stream,
which is probably what one would expect.

BitstreamWriters also flush at __del__-time, so performing:

>>> f = open(file, "wb")
>>> w = BitstreamWriter(f, False)
>>> <do stuff>
>>> f.close()

will generate a warning since the writer will try to flush
a closed stream at garbage-collection time.
Better to call "w.close()" which will perform both the flush and
stream close so that the flush when garbage collected is a no-op.