Add BitstreamQueue as a BitstreamReader subclass
This is designed as a BitstreamReader one can dump data into in bulk
and then process on a bit-by-bit basis without having to dig into
the reader's internals.  It's similar to the buffer-based BitstreamReader
but the buffered one is meant to have a static buffer that isn't changed
after initialization.