Remove DVD-Audio support
I'll resurrect it as a seperate project later (DVD-A Audiotools?)
but without the Python hooks.

I added support for it because I had a few DVD-A discs and thought
it was pretty neat.  But the fact is, it's an obscure, dead format
that takes up a big chunk of the audiotools.decoders module,
isn't expecially portable, and never gets unit tested.