Cleanup bitstream header file
Take a ton of function definitions out of the header file and make them static
so that someone using the bitstream library doesn't have to see them.

However, also make typedefs for the bitstream methods so that
one can pull them out of the struct and use them locally
depending on how the bitstream was opened.