Remove Shorten support
I may go back and re-add Shorten decoding support in the future if
there's demand for it.  I initially added Shorten because the format was
relatively widespread in some circles and the codec seemed relatively
simple.  But the reality is that Shorten is more wonky than it first
appears, it doesn't compress well at all, and doesn't support a lot of
basic features one expects from a lossless codec (like metadata).

My old implementation was actually broken in some respects (like
unsigned 8bps audio) so it's better to omit it for now than leave it in.
This lets me focus more on codecs people actually use.