Update FLAC metadata to handle padding more intelligently
That is, if the old metadata and new metadata's padding blocks
are the same size (as if one gets the metadata, adjusts some fields,
then updates the metadata - like how tracktag does it), adjust
the new metadata's padding block if possible to optimize writing.

But if the padding blocks differ in size (as if one has shrunk
the padding in the new metadata), rewrite the whole file instead.