Replace coverview with coverbrowse
The old coverview utility was limited to a single file at a time.

The new coverbrowse utility allows much easier scanning over
all the audio files in a directory tree without having to open them
individually.

Though in the interest of keeping it small yet portable,
PyGTK support was dropped in favor of using Tkinter only.